Remove the Agama 3 announcement
First of all, this post was never published due to synchronization
problems between github.io and yast.opensuse.org.

The main highlight of Agama 3 was the ability to install Dolomite
Milestone2. But that release is still not ready to be installed via RPMs
(ie. via Agama). So let's remove the post and we will restore it and
adapt it when everything works.